refactor(admin): extract UsersTab into its own file
